+++ test/testckpt.c
@@ -283,7 +283,7 @@ int main (void) {
 	printf ("%s: checkpoint section expiration set\n",
 		get_test_output (error, SA_AIS_OK));
 
-printf ("Please wait, testing expiry of checkpoint sections.\n");
+	printf ("Please wait, testing expiry of checkpoint sections.\n");
 	do {
 	error = saCkptCheckpointRead (checkpointHandle,
 		ReadVectorElements,
@@ -292,7 +292,7 @@ printf ("Please wait, testing expiry of checkpoint sections.\n");
 	} while (error != SA_AIS_ERR_NOT_EXIST);
 	gettimeofday (&tv_end, NULL);
 	timersub (&tv_end, &tv_start, &tv_elapsed);
-	printf ("Elapsed Time to expiry is %ld.%ld (should be about %d seconds)\n", tv_elapsed.tv_sec, tv_elapsed.tv_usec, SECONDS_TO_EXPIRE);
+	printf ("Elapsed Time to expiry is %ld.%ld (should be about %d seconds)\n", tv_elapsed.tv_sec, (unsigned long)tv_elapsed.tv_usec, SECONDS_TO_EXPIRE);
 
 	error = saCkptCheckpointRetentionDurationSet (checkpointHandle,
 						      5000000000LL);
